Refactored e2e tests: check attachment(s) processing
